defineComponent now working
